requirements.txt
matplotlib~=3.7.1
pyyaml~=6.0
tllib~=0.4
torch~=1.13.1+cu116
numpy~=1.23.5
scikit-learn~=1.2.2
tqdm~=4.65.0